Refactor forms endpoint with service and repository
